\name{as.data.frame.details.slma}
\alias{as.data.frame.details.slma}
\title{
Coerce \code{'details.slma'} object to data frame
}
\description{
Coerces a  \code{'details.slma'} object to a data frame.
}
\usage{
\method{as.data.frame}{details.slma}(x, ...)
}
\arguments{
  \item{x}{
   the object \code{x} must be an oject of the class
   \code{"details.slma"}.
  }
  \item{...}{
   the argument \code{...} is not used at the moment.
  }
}
\value{
This function returns a data frame.
}
