\name{as.data.frame.slma}
\alias{as.data.frame.slma}
\title{
Coerce result of SLMA analysis to data frame
}
\description{
Coerces an object of the class \code{slma} to a data frame
}
\usage{
\method{as.data.frame}{slma}(x, ...)
}
\arguments{
  \item{x}{
   the object \code{x} must be an oject of the class \code{"slma"}.
  }
  \item{...}{
   the argument \code{...} is not used at the moment.
  }
}
\value{
This function returns a data frame.
}
