% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc.R
\name{as_conc}
\alias{as_conc}
\title{Coerce data frame to a concordance object}
\usage{
as_conc(x, left = NA, match = NA, right = NA, keep_original = FALSE, ...)
}
\arguments{
\item{x}{A data frame.}

\item{left}{The name of the column in \code{x} that contains the left co-text
of the concordance. Is \code{is.na(left)}, then this column is assumed
to have the name \code{"left"}.}

\item{match}{The name of the column in \code{x} that contains the match
of the concordance. Is \code{is.na(match)}, then this column is assumed
to have the name \code{"match"}.}

\item{right}{The name of the column in \code{x} that contains the right co-text
of the concordance. Is \code{is.na(right)}, then this column is assumed
to have the name \code{"right"}.}

\item{keep_original}{Boolean value. If the values of
\code{left}, \code{match} or \code{right} are not \code{NA}, should
the original names of those columns be kept in the \code{conc} object.}

\item{...}{Additional arguments.}
}
\value{
An object of class \code{conc}.
}
\description{
Coerce a data frame to an object of the class \code{conc}.
}
\examples{
(conc_data <- conc('A very small corpus.', '\\\\w+', as_text = TRUE))
df <- as.data.frame(conc_data)
as_conc(df)
}
