% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnames.R
\name{as_fnames}
\alias{as_fnames}
\title{Coerce object to 'fnames'}
\usage{
as_fnames(x, remove_duplicates = TRUE, sort = TRUE, ...)
}
\arguments{
\item{x}{A character vector (or a \code{\link{freqlist}} object!)}

\item{remove_duplicates}{Boolean. Whether duplicates should be removed.}

\item{sort}{Boolean. Whether the output should be sorted.}

\item{...}{Additional arguments.}
}
\value{
An object of class \code{\link{fnames}}.
}
\description{
This function coerces a character vector into an object of class \code{\link{fnames}}.
}
\examples{
as_fnames("path/to/my/corpus_file")
}
