% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc.R
\name{explore.conc}
\alias{explore.conc}
\title{Interactively navigate through concordance list}
\usage{
\method{explore}{conc}(x, n = 20, from = 1, use_clear = TRUE, ...)
}
\arguments{
\item{x}{Object of class \code{conc}.}

\item{n}{Maximum number of items to show at once.}

\item{from}{Index of the first item to be shown in the console at the beginning.}

\item{use_clear}{Boolean value. If \code{TRUE} and the feature is supported by
the R environment, the console will be cleared in between all the interactive
steps in the exploration session.}

\item{...}{Additional arguments.}
}
\value{
Invisibly, \code{x}.
}
\description{
This method only works in an interactive R session to open
'exploration mode', in which the user can navigate through the \code{conc}
object \code{x} by means of brief commands. In 'exploration mode' the user can
ask of a list of available commands by keying in \code{?}, followed by ENTER.
The user can quiet 'exploration mode' by keying in \code{q}, followed by ENTER.
}
