% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/assoc.R, R/conc.R, R/fnames.R,
%   R/freqlist.R, R/types.R
\name{explore}
\alias{explore}
\alias{explore.assoc_scores}
\alias{explore.conc}
\alias{explore.fnames}
\alias{explore.freqlist}
\alias{explore.default}
\alias{explore.types}
\title{Interactively navigate through an object}
\usage{
explore(x, ...)

\method{explore}{assoc_scores}(
  x,
  n = 20,
  from = 1,
  from_col = 1,
  perl = TRUE,
  sort_order = c("none", "G_signed", "PMI", "alpha"),
  use_clear = TRUE,
  ...
)

\method{explore}{conc}(x, n = 20, from = 1, use_clear = TRUE, ...)

\method{explore}{fnames}(x, n = 20, from = 1, perl = TRUE, use_clear = TRUE, ...)

\method{explore}{freqlist}(x, n = 20, from = 1, perl = TRUE, use_clear = TRUE, ...)

\method{explore}{default}(x, ...)

\method{explore}{types}(x, n = 20, from = 1, perl = TRUE, use_clear = TRUE, ...)
}
\arguments{
\item{x}{An object of any of the classes for which the method is implemented.}

\item{...}{Additional arguments.}

\item{n}{Maximum number of items in the object to be printed at once.}

\item{from}{Index of the first item to be printed.}

\item{from_col}{Index of the first column to be displayed in the regular area
(among all selected columns, including frozen columns). If \code{from_col} points}

\item{perl}{Boolean value. Whether or not the regular expressions used in the
exploration session use the PERL flavour of regular expression.}

\item{sort_order}{Order in which the items are to be printed. In general, possible values
are \code{"alpha"} (meaning that the items are to be sorted alphabetically),
and \code{"none"} (meaning that the items are not to be sorted).
If \code{x} is an object of class \code{\link{assoc_scores}}, a column name
or vector of column names may be provided instead.}

\item{use_clear}{Boolean. If \code{TRUE}, and if the feature is supported by the R
environment, the console will be cleared in between all interactive steps
in the exploration session.}
}
\value{
Invisibly, \code{x}.
}
\description{
This method only works in an interactive R session to open
'exploration mode', in which the user can navigate through the
object \code{x} by means of brief commands. In 'exploration mode' the user can
ask of a list of available commands by keying in \verb{?}, followed by \code{ENTER}.
The user can quiet 'exploration mode' by keying in \code{q}, followed by \code{ENTER}.
}
