% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{explore.types}
\alias{explore.types}
\title{Interactively navigate through 'types' object}
\usage{
\method{explore}{types}(x, n = 20, from = 1, perl = TRUE, use_clear = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{types}.}

\item{n}{Maximum number of items in the \code{types} object to be shown at once.}

\item{from}{First item in the types object that is shown at the beginning of
the explroation session.}

\item{perl}{Boolean. Whether or not regular expressions used in the exploration
session use the PERL flavor of regular expression.}

\item{use_clear}{Boolean value. If \code{use_clear} is \code{TRUE},
and if moreover the feature is supported by the R environment,
the console will be cleared in between all interactive steps in the exploration session.}

\item{...}{Additional arguments (not implemented)}
}
\value{
Invisibly, \code{x}
}
\description{
This method only works in an interactive R session to open
  'exploration mode', in which the user can navigate through the \code{types}
  object \code{x} by means of brief commands. In 'exploration mode' the user can
  ask of a list of available commands by keying in \code{?}, followed by ENTER.
  The user can quiet 'exploration mode' by keying in \code{q}, followed by ENTER.
}
