% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnames.R
\name{fnames}
\alias{fnames}
\alias{get_fnames}
\title{Retrieve the names of files in a given path}
\usage{
get_fnames(
  path = ".",
  re_pattern = NULL,
  recursive = TRUE,
  perl = TRUE,
  invert = FALSE
)
}
\arguments{
\item{path}{The location of the files to be listed.}

\item{re_pattern}{Optional regular expression. If present, then only the
filenames that match it are retrieved (unless \code{invert = TRUE}, in which
case those filenames are excluded). The match is done over the absolute
path of the files.}

\item{recursive}{Boolean value. Should the subdirectories of \code{path} also be
searched?}

\item{perl}{Boolean value. Whether \code{re_pattern} should be interpreted as a
PERL flavour of regular expression.}

\item{invert}{Boolean value. If \code{TRUE}, filenames matching \code{re_pattern} are
the only ones retrieved. If \code{FALSE}, filenames matching \code{re_pattern} are
excluded.}
}
\value{
An object of class \code{fnames}, which is a special kind of character
vector storing the absolute paths of the corpus files.
}
\description{
Build an object of class \code{fnames}.
}
\examples{
\dontrun{
cwd_fnames <- get_fnames(recursive = FALSE)
cwd_fnames
print(cwd_fnames)
as.data.frame(cwd_fnames)
as_tibble(cwd_fnames)

sort(cwd_fnames)

summary(cwd_fnames)
}
}
