\name{mclm-classes}
\alias{mclm-classes}
\title{
Main Classes in the "mclm" Package
}
\description{
Description of the main classes in the "mclm" package.
}
\details{
  The class \code{'tokens'} is used to represent \emph{token sequences}. 

  The class \code{'types'} is used to represent \emph{type collections}. 

  The class \code{'freqlist'} is used to represent \emph{frequency list}.
  The term "list" in its name is used informally. It is not intended to
  indicate that this class is a list in R's technical sense.
  
  The class \code{'assoc_scores'} is used to represent
  \emph{association scores}.

  The class \code{'df'} is used to represent \emph{data frames}. 


}