\name{mclm-package}
\alias{mclm-package}
\alias{mclm}
\docType{package}
\title{
Mastering Corpus Linguistics Methods
}
\description{
A collection of functions in support of quantitative corpus linguistics.
}
\details{
\tabular{ll}{
Package: \tab mclm\cr
Version: \tab 0.2\cr
Date: \tab 2017-10-13\cr
Suggests: \tab MASS\cr
Depends: \tab ca, tm, XML\cr
License: \tab GPL (>=2)\cr
}
% ~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Dirk Speelman

Maintainer: Dirk Speelman <Dirk.Speelman@kuleuven.be>
}
\references{
This package accompanies Speelman, D. 2018. \emph{Mastering Corpus Linguistics Methods:
 A Practical Introduction with AntConc and R}, John Wiley & Sons.
}
\keyword{ package }
%\seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%
%}
