% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqlist.R
\name{n_freqlist}
\alias{n_freqlist}
\alias{n_tokens.freqlist}
\alias{n_types.freqlist}
\title{Count types or tokens in a 'freqlist' Object}
\usage{
\method{n_tokens}{freqlist}(x, ...)

\method{n_types}{freqlist}(x, ...)
}
\arguments{
\item{x}{Object of class \code{freqlist}.}

\item{...}{Additional arguments.}
}
\value{
A number.
}
\description{
\code{n_tokens} returns the number of tokens in \code{x}.
\code{n_types} returns the number of tokens in \code{x}.
}
\section{Functions}{
\itemize{
\item \code{n_tokens.freqlist}: Count number of tokens

\item \code{n_types.freqlist}: Count number of types
}}

\examples{
(tks <- tokenize("The old man and the sea."))
n_tokens(tks)

(flist <- freqlist(tks))
n_tokens(flist)
n_types(flist)
}
