% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{p_to_chisq1}
\alias{p_to_chisq1}
\title{P right quantile in chi-squared distribution with 1 df}
\usage{
p_to_chisq1(p)
}
\arguments{
\item{p}{A proportion.}
}
\value{
The \code{p} \emph{right quantile} in the \eqn{\chi^2} distribution with
one degree of freedom.
}
\description{
P right quantile that takes as its argument a probability \code{p} and that returns
the \code{p} \emph{right quantile} in the \eqn{\chi^2} distribution with one degree of
freedom. In other words, it returns a value \emph{q} such that a proportion \code{p}
\eqn{\chi^2} distribution with one degree of freedom lies above \emph{q}.
}
\seealso{
\code{\link[=chisq1_to_p]{chisq1_to_p()}}
}
