\name{print.details.slma}
\alias{print.details.slma}
\title{
Print detailed SLMA information for a specific linguistic item
}
\description{
Print objects of the class \code{"details.slma"}. 
}
\usage{
\method{print}{details.slma}(x, 
      ...)
}
\arguments{
  \item{x}{
  an object of the class \code{"details.slma"}
  }
  \item{...}{
  additional printing arguments
  }
}
\value{
Invisibly returns \code{x}.
}
