\name{print.freqlist}
\alias{print.freqlist}
\title{
Print a frequency list
}
\description{
Print objects of the class \code{"freqlist"}. 
}
\usage{
\method{print}{freqlist}(x,
                         n = 6, 
                         ...)
}
\arguments{
  \item{x}{
  a frequency list
  }
  \item{n}{
  number of top frequency items to print
  }
  \item{...}{
  additional printing arguments
  }
}
\value{
Invisibly returns \code{x}.
}
