\name{print.slma}
\alias{print.slma}
\title{
Prints the result of an SLMA analysis.
}
\description{
Print objects of the class \code{"slma"}. 
}
\usage{
\method{print}{slma}(x,
      n = 20,
      from = 1,
      ...)
}
\arguments{
  \item{x}{
  the result of an SLMA analysis.
  }
  \item{n}{
  number of items of the analysis to print.
  }
  \item{from}{
  first item to print.
  }
  \item{...}{
  additional printing arguments
  }
}
\value{
Invisibly returns \code{x}.
}
