\name{print.zoom_re.freqlist}
\alias{print.zoom_re.freqlist}
\title{
Print frequency list, with `zoom' turned on.
}
\description{
Print objects of the class \code{"zoom_re.freqlist"}. 
}
\usage{
\method{print}{zoom_re.freqlist}(x,
                                 n = 6, 
                                 ...)
}
\arguments{
  \item{x}{
  a frequency list
  }
  \item{n}{
  number of top frequency items to print
  }
  \item{...}{
  additional printing arguments
  }
}
\value{
Invisibly returns \code{x}.
}
