% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{sort.types}
\alias{sort.types}
\title{Sort a collection of types}
\usage{
\method{sort}{types}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{An object of the class \code{types}.}

\item{decreasing}{Logical value that indicates whether the items should be
sorted from small to large (when \code{decreasing} in \code{FALSE})
or from large to small (when \code{decreasing} is \code{TRUE}).
The default value is \code{FALSE}.}

\item{...}{Other arguments}
}
\value{
An object of class \code{types}.
}
\description{
Sort an object of the class \code{types}.
}
\details{
At the moment, types collections are not allowed to contain \code{NA} values.
  Therefore, no function argument is available
  that specifies how \code{NA} values should be sorted.
}
\examples{
(tps <- as_types(c("the", "a", "some", "no")))
sort(tps)
sort(tps, decreasing = TRUE)
}
