% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqlist.R
\name{summary.freqlist}
\alias{summary.freqlist}
\alias{print.summary.freqlist}
\title{Succint description of a 'freqlist' object}
\usage{
\method{summary}{freqlist}(object, ...)

\method{print}{summary.freqlist}(x, ...)
}
\arguments{
\item{object}{Object of class \code{freqlist}.}

\item{...}{Additional arguments}

\item{x}{Object of class \code{summary.freqlist}.}
}
\value{
Object of class \code{summary.freqlist}
}
\description{
Build and/or print an object of the class \code{summary.freqlist}.
}
\examples{
flist <- freqlist(tokenize("the old man and the sea."))
summary(flist) 
print(summary(flist))

(flist_sum <- summary(flist))
names(flist_sum)
flist_sum[["n_tokens"]]
flist_sum$n_tokens
flist_sum[["n_types"]]
flist_sum$n_types

flist2 <- flist["man", invert = TRUE]
(flist2_sum <- summary(flist2))
names(flist2_sum)
flist2_sum[["n_tokens"]]
flist2_sum$n_tokens
flist2_sum[["n_types"]]
flist2_sum$n_types
flist2_sum[["tot_n_tokens"]]
flist2_sum$tot_n_tokens
}
