% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqlist.R
\name{tot_n_tokens.freqlist}
\alias{tot_n_tokens.freqlist}
\alias{tot_n_tokens<-.freqlist}
\title{Retrieve or set the total number of tokens}
\usage{
\method{tot_n_tokens}{freqlist}(x) <- value

\method{tot_n_tokens}{freqlist}(x)
}
\arguments{
\item{x}{Object of class \code{freqlist}.}

\item{value}{Numerical value.}
}
\value{
Numerical value.
}
\description{
These methods retrieve or set, for a \code{freqlist}, the total number of tokens in
the corpus on which the frequency counts are based.
This total number of tokens may be higher than the sum of all frequency
counts in \code{x}, for instance, if \code{x} contains frequency counts
for a selection of items only, and not for all tokens in the corpus.
}
\section{Functions}{
\itemize{
\item \code{tot_n_tokens<-.freqlist}: Set total number of tokens

\item \code{tot_n_tokens.freqlist}: Retrieve the total number of tokens
}}

\examples{
x <- freqlist("The man and the mouse.",
              re_token_splitter = "(?xi) [:\\\\s.;,?!\"]+",
              as_text = TRUE)
x
tot_n_tokens(x)

y <- keep_types(x, c("man", "and"))
tot_n_tokens(y)
y

tot_n_tokens(y) <- sum(y)
y
tot_n_tokens(y)
}
\seealso{
Other freqlist setters and getters: 
\code{\link{orig_ranks.freqlist}()},
\code{\link{ranks.freqlist}()}
}
\concept{freqlist setters and getters}
