% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc.R
\name{as.data.frame.conc}
\alias{as.data.frame.conc}
\title{Coerce concordance to data frame}
\usage{
\method{as.data.frame}{conc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{conc}.}

\item{...}{Additional arguments.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Coerce object of class \code{conc} to \code{data.frame}.
}
\examples{
(conc_data <- conc('A very small corpus.', '\\\\w+', as_text = TRUE))
as.data.frame(conc_data)
}
\seealso{
\code{\link{as_tibble.conc}}
}
