% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqlist.R
\name{as.data.frame.freqlist}
\alias{as.data.frame.freqlist}
\title{Coerce frequency list to data frame}
\usage{
\method{as.data.frame}{freqlist}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{freqlist}.}

\item{row.names}{\code{NULL} or a character vector giving the row names for 
the data frame. Missing values are not allowed.}

\item{optional}{Boolean. Whether setting row names and converting column names
to syntactic names is optional. See \code{\link{as.data.frame}}.}

\item{...}{Additional arguments.}
}
\value{
A \code{data.frame}.
}
\description{
This method coerces a \code{freqlist} to a \code{data.frame}.
}
\examples{
toy_corpus <- "Once upon a time there was a tiny toy corpus.
It consisted of three sentences. And it lived happily ever after."

(flist <- freqlist(toy_corpus, as_text = TRUE))
as.data.frame(flist)

(flist2 <- keep_re(flist, "^..?$"))
as.data.frame(flist2)
}
\seealso{
\code{\link{as_tibble.freqlist}}
}
