% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{as.data.frame.types}
\alias{as.data.frame.types}
\title{Coerce `types` object to Data Frame}
\usage{
\method{as.data.frame}{types}(x, ...)
}
\arguments{
\item{x}{An object of class \code{types}.}

\item{...}{Other arguments.}
}
\value{
A dataframe. In this dataframe, the types sit in a column named \code{type}.
}
\description{
The S3 method \code{as.data.frame}, when applied to an object \code{x} of the
  class \code{types}, coerces an object of the class \code{types} to a data frame.
}
\examples{
toy_corpus <- "Once upon a time there was a tiny toy corpus.
It consisted of three sentence. And it lived happily ever after."

(typs <- types(toy_corpus, as_text = TRUE))

as.data.frame(typs)
}
