% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc.R
\name{as_tibble.conc}
\alias{as_tibble.conc}
\title{Coerce concordance to tibble}
\usage{
\method{as_tibble}{conc}(x, ...)
}
\arguments{
\item{x}{Object of class \code{conc}.}

\item{...}{Additional arguments.}
}
\value{
Object of class \code{\link[tibble]{tibble}}.
}
\description{
Coerce an object of class \code{conc} to \code{\link[tibble]{tibble}}.
}
\examples{
(conc_data <- conc('A very small corpus.', '\\\\w+', as_text = TRUE))
as_tibble(conc_data)
}
