% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqlist.R
\name{as_tibble.freqlist}
\alias{as_tibble.freqlist}
\title{Coerce frequency list into a tibble}
\usage{
\method{as_tibble}{freqlist}(x, ...)
}
\arguments{
\item{x}{Object of class \code{freqlist}.}

\item{...}{Additional arguments.}
}
\value{
Object of class \code{\link[tibble]{tibble}}.
}
\description{
This method coerces \code{freqlist} into \code{\link[tibble]{tibble}}.
}
\examples{
toy_corpus <- "Once upon a time there was a tiny toy corpus.
It consisted of three sentences. And it lived happily ever after."

(flist <- freqlist(toy_corpus, as_text = TRUE))
as_tibble(flist)

(flist2 <- keep_re(flist, "^..?$"))
as_tibble(flist2)
}
\seealso{
\code{\link{as.data.frame.freqlist}}
}
