% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{as_tibble.types}
\alias{as_tibble.types}
\title{Coerce 'types' object to tibble}
\usage{
\method{as_tibble}{types}(x, ...)
}
\arguments{
\item{x}{An object of the class \code{types}.}

\item{...}{Other arguments}
}
\value{
A tibble. In this tibble, the types sit in a colum named \code{type}.
}
\description{
The S3 method \code{as_tibble}, when applied to an object \code{x} of the
  class \code{types}, coerces an object of the class \code{types} to a tibble.
}
\examples{
toy_corpus <- "Once upon a time there was a tiny toy corpus.
It consisted of three sentence. And it lived happily ever after."

(typs <- types(toy_corpus, as_text = TRUE))

as_tibble(typs)
}
\seealso{
as.data.frame.types
}
