% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{as_types}
\alias{as_types}
\title{Coerce object to a vector of types}
\usage{
as_types(x, remove_duplicates = TRUE, sort = TRUE, ...)
}
\arguments{
\item{x}{Object to coerce}

\item{remove_duplicates}{Length one boolean vector that determines whether or not
duplicates are removed from \code{x} prior to coercing to a vector of types.}

\item{sort}{Length one boolean vector that determines whether or not
\code{x} is alphabetically sorted prior to coercing to a vector of types;
this argument is ignored if \code{remove_duplicates} is \code{TRUE},
because the result of removing duplicates is always sorted.}

\item{...}{Additional arguments (not implemented)}
}
\value{
An object of class \code{types}.
}
\description{
This function coerces an object, such as a character vector, to an object of
class \code{types}.
}
\examples{

toy_corpus <- "Once upon a time there was a tiny toy corpus.
It consisted of three sentences. And it lived happily ever after."

flist <- freqlist(toy_corpus, re_token_splitter = "\\\\\\\\W+", as_text = TRUE)
print(flist, n = 1000)
(sel_types <- as_types(c("happily", "lived", "once")))
keep_types(flist, sel_types)
tks <- tokenize(toy_corpus, re_token_splitter = "\\\\\\\\W+")
print(tks, n = 1000)
tks[3:12] # idx is relative to selection
head(tks) # idx is relative to selection
tail(tks) # idx is relative to selection
}
