% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{chisq1_to_p}
\alias{chisq1_to_p}
\title{Proportion of chi-squared distribution with 1 df that sits to the right of x
Helper function that takes as its argument a numerical value \code{x} and
that returns the proportion \emph{p} of the chi-squared
distribution with one degree of freedom that sits to the right of the
value `x.}
\usage{
chisq1_to_p(x)
}
\arguments{
\item{x}{A number.}
}
\value{
The proportion \emph{p} of the chi-squared distribution with one
degree of freedom that sits to the right of the value \code{x}.
}
\description{
Proportion of chi-squared distribution with 1 df that sits to the right of x
Helper function that takes as its argument a numerical value \code{x} and
that returns the proportion \emph{p} of the chi-squared
distribution with one degree of freedom that sits to the right of the
value `x.
}
\seealso{
\code{\link[=p_to_chisq1]{p_to_chisq1()}}
}
