% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/assoc.R, R/conc.R, R/freqlist.R,
%   R/types.R
\name{explore}
\alias{explore}
\alias{explore.assoc_scores}
\alias{explore.conc}
\alias{explore.freqlist}
\alias{explore.types}
\title{Interactively navigate through an object}
\usage{
explore(x, ...)

\method{explore}{assoc_scores}(
  x,
  n = 20,
  from = 1,
  from_col = 1,
  perl = TRUE,
  sort_order = c("none", "G_signed", "PMI", "alpha"),
  use_clear = TRUE,
  ...
)

\method{explore}{conc}(x, n = 20, from = 1, use_clear = TRUE, ...)

\method{explore}{freqlist}(x, n = 20, from = 1, perl = TRUE, use_clear = TRUE, ...)

\method{explore}{types}(x, n = 20, from = 1, perl = TRUE, use_clear = TRUE, ...)
}
\arguments{
\item{x}{An object of any of the classes for which the method is implemented.}

\item{...}{Additional arguments.}

\item{n}{Maximum number of items in the object to be shown at once.}

\item{from}{Index of the first item to be shown at the beginning of the
interactive session.}

\item{from_col}{Index of the first column to be displayed (among all selected
columns, including frozen columns).}

\item{perl}{Boolean value. Whether or not the regular expressions used in the
exploration session use the PERL flavour of regular expression.}

\item{sort_order}{Order in which the items are to be printed. If \code{"alpha"} is
given, the items will be sorted alphabetically; \code{"none"} indicates that no
sorting should be done. A column name can also be provided.}

\item{use_clear}{Boolean. If \code{TRUE}, and if the feature is supported by the R
environment, the console will be cleared in between all interactive steps
in the exploration session.}
}
\value{
Invisibly, \code{x}.
}
\description{
This method only works in an interactive R session to open
'exploration mode', in which the user can navigate through the
object \code{x} by means of brief commands. In 'exploration mode' the user can
ask of a list of available commands by keying in \verb{?}, followed by \code{ENTER}.
The user can quiet 'exploration mode' by keying in \code{q}, followed by \code{ENTER}.
}
