% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{find_xpath}
\alias{find_xpath}
\title{Run XPath query}
\usage{
find_xpath(
  pattern,
  x,
  handlers = NULL,
  trim = TRUE,
  fun = NULL,
  final_fun = NULL,
  namespaces = NULL,
  ...
)
}
\arguments{
\item{pattern}{An XPath query.}

\item{x}{A corpus: an \code{\link{fnames}} object, a character vector of an XML source,
a list of documents already parsed with \code{\link[XML:xmlTreeParse]{XML::xmlParse()}}.}

\item{handlers}{Optional collection of functions used to map the different
XML nodes to R objects.}

\item{trim}{Logical. Whether to strip of whitespace from the begginng and end
of text strings.}

\item{fun}{Optional function to be applied to the individual matches prior
to returning the result.}

\item{final_fun}{Optional final function to be applied to the complete list
of matches prior to returning the result.}

\item{namespaces}{Optional namespaces.}

\item{...}{Additional arguments.}
}
\value{
A list with matches (see \code{\link[XML:getNodeSet]{XML::xpathApply()}}).
}
\description{
This function finds matches for an XPath query in a corpus.
}
\examples{
test_xml <- '
<p>
  <w pos="at">The</w>
  <w pos="nn">example</w>
  <punct>.</punct>
</p>'

find_xpath("//w", test_xml)
find_xpath("@pos", test_xml)
find_xpath("//w[@pos='nn']", test_xml)
}
