% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnames.R
\name{fnames_merge}
\alias{fnames_merge}
\alias{fnames_merge_all}
\title{Merge filenames collections}
\usage{
fnames_merge(x, y, sort = FALSE)

fnames_merge_all(..., sort = FALSE)
}
\arguments{
\item{x, y}{An object of class \code{\link{fnames}}.}

\item{sort}{Boolean value. Should the items in the output be sorted?}

\item{...}{Various objects of class \code{\link{fnames}} or a list of
objects of class \code{\link{fnames}}.}
}
\value{
An object of class \code{\link{fnames}}.
}
\description{
These functions merge two or more \code{\link{fnames}} objects into one larger \code{\link{fnames}}
object, removing duplicates (keeping only the first appearance) and only
resorting the items if \code{sort = TRUE}.
}
\examples{
\dontrun{
cwd_fnames <- get_fnames(recursive = FALSE)
cwd_fnames2 <- get_fnames("some_subdirectory")
cwd_fnames3 <- get_fnames("another_subdirectory")
fnames_merge(cwd_fnames, cwd_fnames2)
fnames_merge_all(cwd_fnames, cwd_fnames2, cwd_fnames3)
}
}
