% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqlist.R
\name{freqlist_diff}
\alias{freqlist_diff}
\title{Subtract frequency lists}
\usage{
freqlist_diff(x, y)
}
\arguments{
\item{x, y}{Objects of class \code{freqlist}.}
}
\value{
An object of class \code{freqlist}.
}
\description{
This function merges information from two frequency lists, subtracting the frequencies found
in the second frequency lists from the frequencies found in the first list.
}
\examples{
(flist1 <- freqlist("A first toy corpus.", as_text = TRUE))
(flist2 <- freqlist("A second toy corpus.", as_text = TRUE))

freqlist_diff(flist1, flist2)
}
