% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc.R
\name{import_conc}
\alias{import_conc}
\title{Import a concordance}
\usage{
import_conc(x, file_encoding = "UTF-8", source_type = c("corpuseye"), ...)
}
\arguments{
\item{x}{A vector of input filenames.}

\item{file_encoding}{Encoding of the file(s).}

\item{source_type}{Character string. How the file is read. Currently only
\code{"corpuseye"} is supported.}

\item{...}{Additional arguments (not implemented).}
}
\value{
An object of class \code{\link{conc}}.
}
\description{
This function imports a concordance from files generated by other means than
\code{\link[=write_conc]{write_conc()}}.
}
\seealso{
\code{\link[=read_conc]{read_conc()}} for files written with \code{\link[=write_conc]{write_conc()}}.
}
