% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclm-package.R
\docType{package}
\name{mclm-package}
\alias{mclm}
\alias{mclm-package}
\title{Mastering Corpus Linguistic Methods}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Read, inspect and process corpus files for quantitative corpus linguistics. Obtain concordances via regular expressions, tokenize texts, and compute frequencies and association measures. Useful for collocation analysis, keywords analysis and variationist studies (comparison of linguistic variants and of linguistic varieties).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/masterclm/mclm}
  \item \url{https://masterclm.github.io/mclm/}
}

}
\author{
\strong{Maintainer}: Mariana Montes \email{mariana.montes@kuleuven.be}

Authors:
\itemize{
  \item Dirk Speelman \email{dirk.speelman@kuleuven.be}
}

}
\keyword{internal}
