% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclm-package.R
\docType{package}
\name{mclm-package}
\alias{mclm}
\alias{mclm-package}
\title{Mastering Corpus Linguistic Methods}
\description{
Read corpus files, create frequency lists, generate concordances, compute co-occurrence
frequencies and perform keyword analysis, stable lexical markers analysis
and more. Companion to the Methods in Corpus Linguistics course at the Advanced
Master in Linguistics (KU Leuven).
}
\author{
\strong{Maintainer}: Dirk Speelman \email{dirk.speelman@kuleuven.be}

Authors:
\itemize{
  \item Mariana Montes \email{mariana.montes@kuleuven.be}
}

}
\keyword{internal}
