% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclm-package.R
\docType{package}
\name{mclm-package}
\alias{mclm}
\alias{mclm-package}
\title{Mastering Corpus Linguistic Methods}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Various functions in support of quantitative corpus linguistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://montesmariana.github.io/mclm}
  \item \url{https://github.com/montesmariana/mclm}
  \item \url{https://montesmariana.github.io/mclm/}
}

}
\author{
\strong{Maintainer}: Mariana Montes \email{mariana.montes@kuleuven.be}

Authors:
\itemize{
  \item Dirk Speelman \email{dirk.speelman@kuleuven.be}
}

}
\keyword{internal}
