\name{mclm-package}
\alias{mclm-package}
\alias{mclm}
\docType{package}
\title{
Mastering Corpus Linguistics Methods
}
\description{
A collection of functions in support of quantitative corpus linguistics.
}
\details{
\tabular{ll}{
Package: \tab mclm\cr
Version: \tab 0.2.5\cr
Date: \tab 2021-10-08\cr
Depends: \tab ca, tibble\cr
Imports: \tab crayon, dplyr, Rcpp, readr, stringi, stringr, tm, XML, yaml\cr
Suggests: \tab MASS\cr
License: \tab GPL (>=2)\cr
}
% ~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Dirk Speelman

Maintainer: Dirk Speelman <Dirk.Speelman@kuleuven.be>
}
\references{
See: \url{https://wwwling.arts.kuleuven.be/mclm/}
}
\keyword{ package }
%\seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%
%}
