% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnames.R
\name{n_fnames}
\alias{n_fnames}
\title{Count number of items in an 'fnames' object}
\usage{
n_fnames(x, ...)
}
\arguments{
\item{x}{Object of class \code{\link{fnames}}.}

\item{...}{Additional arguments.}
}
\value{
A number.
}
\description{
This function counts the number of items, duplicated or not, in an \code{\link{fnames}}
object. If there are duplicated items, it will return a warning.
}
\examples{
\dontrun{
cwd_fnames <- get_fnames()
n_fnames(cwd_fnames)
}
}
