% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqlist.R
\name{n_tokens.freqlist}
\alias{n_tokens.freqlist}
\title{Give Number of Tokens in a 'freqlist' Object}
\usage{
\method{n_tokens}{freqlist}(x, ...)
}
\arguments{
\item{x}{Object of class \code{freqlist}.}

\item{...}{Additional arguments.}
}
\value{
A number.
}
\description{
Return the number of tokens in \code{x}.
}
\examples{
(tks <- tokenize("The old man and the sea."))
n_tokens(tks)

(flist <- freqlist(tks))
n_tokens(flist)
n_types(flist)
}
