% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{n_types.types}
\alias{n_types.types}
\title{Give number of types in a 'types' object}
\usage{
\method{n_types}{types}(x, ...)
}
\arguments{
\item{x}{An object of the class \code{types}.}

\item{...}{Additional arguments}
}
\value{
The number of types in \code{x}.
}
\description{
When applied to an object of class \code{types}, it returns the number of types.
}
\examples{
(tks <- tokenize("The old man and the sea."))
n_tokens(tks)
(tps <- types(tks))
n_types(tps)
}
