% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqlist.R
\name{orig_ranks.freqlist}
\alias{orig_ranks.freqlist}
\alias{orig_ranks<-.freqlist}
\title{Retrieve or set original ranks}
\usage{
\method{orig_ranks}{freqlist}(x) <- value

\method{orig_ranks}{freqlist}(x, with_names = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{freqlist}.}

\item{value}{Currently it can only be \code{NULL}.}

\item{with_names}{Boolean. Whether or not the items in the output should
be given names. If \code{TRUE}, then the names
of the types in the frequency list are used as names.}

\item{...}{Additional arguments.}
}
\value{
Either \code{NULL} or a numeric vector, representing the
  original ranks, with as its names the types to which these ranks apply.
}
\description{
These methods retrieve or set, for a \code{freqlist}, the original ranks for its frequency
counts.
These original ranks are only defined if \code{x} is the result of a selection
procedure (i.e. if \code{x} contains frequency counts for a selection of items
only, and not for all tokens in the corpus).
}
\section{Functions}{
\itemize{
\item \code{orig_ranks<-.freqlist}: Set original ranks

\item \code{orig_ranks.freqlist}: Get original ranks
}}

\examples{
x <- freqlist("The man and the mouse.",
              as_text = TRUE)
x
orig_ranks(x)
orig_ranks(x, with_names = TRUE)

y <- keep_types(x, c("man", "and"))
orig_ranks(y)
y

orig_ranks(y) <- NULL
y
orig_ranks(y)

tot_n_tokens(y) <- sum(y)
y
}
\seealso{
Other freqlist setters and getters: 
\code{\link{ranks.freqlist}()},
\code{\link{tot_n_tokens.freqlist}()}
}
\concept{freqlist setters and getters}
