% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc.R
\name{print.conc}
\alias{print.conc}
\title{Print a concordance-based data frame.}
\usage{
\method{print}{conc}(x, n = 30, ...)
}
\arguments{
\item{x}{An object of class \code{conc}.}

\item{n}{Number of items to print in the KWIC display.}

\item{...}{Additional arguments for \code{\link{print_kwic}}.}
}
\value{
Invisibly, \code{x}.
}
\description{
Print an object of class \code{conc}.
}
\examples{
(conc_data <- conc('A very small corpus.', '\\\\w+', as_text = TRUE))
print_kwic(conc_data)
print(conc_data, n = 3)
}
