% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqlist.R
\name{print.freqlist}
\alias{print.freqlist}
\title{Print a frequency list}
\usage{
\method{print}{freqlist}(x, n = 20, from = 1, extra = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{freqlist}.}

\item{n}{Number of items to print.}

\item{from}{Position of the first item to print.}

\item{extra}{Extra settings.}

\item{...}{Additional arguments.}
}
\value{
Invisibly, \code{x}.
}
\description{
Print an object of class \code{freqlist}.
}
\examples{
toy_corpus <- "Once upon a time there was a tiny toy corpus.
It consisted of three sentences. And it lived happily ever after."
(flist <- freqlist(toy_corpus, as_text = TRUE))
print(flist)
print(flist, from = 7, n = 10)
}
