% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{print.types}
\alias{print.types}
\title{Print a vector of 'types'}
\usage{
\method{print}{types}(x, n = 20, from = 1, sort_order = c("none", "alpha"), extra = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{types}.}

\item{n}{Maximum number of types to print.}

\item{from}{Position of the first item to print.}

\item{sort_order}{Order in which the items are to be printed. Possible value
are \code{"alpha"} (meaning that the items are to be sorted alphabetically),
and \code{"none"} (meaning that the items are not to be sorted).}

\item{extra}{Extra settings.}

\item{...}{Additional printing arguments.}
}
\value{
Invisibly, \code{x}.
}
\description{
Print objects of the class \code{types}.
}
\examples{
x <- as_types(c("first", "second", "third"))
print(x, n = 1000)
}
