% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnames.R, R/freqlist.R, R/generics.R, R/re.R,
%   R/stubs.R, R/types.R
\name{plot.fnames}
\alias{plot.fnames}
\alias{plot.summary.fnames}
\alias{[<-.freqlist}
\alias{plot.freqlist}
\alias{plot.summary.freqlist}
\alias{explore.default}
\alias{keep_pos.default}
\alias{drop_pos.default}
\alias{keep_re.default}
\alias{drop_re.default}
\alias{keep_types.default}
\alias{drop_types.default}
\alias{keep_bool.default}
\alias{drop_bool.default}
\alias{n_tokens.default}
\alias{n_types.default}
\alias{type_names.default}
\alias{tot_n_tokens.default}
\alias{tot_n_tokens<-.default}
\alias{orig_ranks.default}
\alias{ranks.default}
\alias{plot.re}
\alias{plot.summary.re}
\alias{stubs}
\alias{plot.types}
\alias{plot.summary.types}
\title{Stubs (Functions/Methods Not Implemented)}
\usage{
\method{plot}{fnames}(x, ...)

\method{plot}{summary.fnames}(x, ...)

\method{[}{freqlist}(x, i, ...) <- value

\method{plot}{freqlist}(x, ...)

\method{plot}{summary.freqlist}(x, ...)

\method{explore}{default}(x, ...)

\method{keep_pos}{default}(x, pos, invert = FALSE, ...)

\method{drop_pos}{default}(x, pos, ...)

\method{keep_re}{default}(x, pattern, perl = TRUE, invert = FALSE, ...)

\method{drop_re}{default}(x, pattern, perl = TRUE, ...)

\method{keep_types}{default}(x, types, invert = FALSE, ...)

\method{drop_types}{default}(x, types, ...)

\method{keep_bool}{default}(x, bool, invert = FALSE, ...)

\method{drop_bool}{default}(x, bool, ...)

\method{n_tokens}{default}(x, ...)

\method{n_types}{default}(x, ...)

\method{type_names}{default}(x, ...)

\method{tot_n_tokens}{default}(x)

\method{tot_n_tokens}{default}(x) <- value

\method{orig_ranks}{default}(x, ...)

\method{ranks}{default}(x, ...)

\method{plot}{re}(x, ...)

\method{plot}{summary.re}(x, ...)

\method{plot}{types}(x, ...)

\method{plot}{summary.types}(x, ...)
}
\arguments{
\item{x}{Some object}

\item{...}{Additional arguments}

\item{i}{Index}

\item{value}{Value to change}

\item{invert}{Boolean value.}
}
\value{
Invisibly, \code{NULL}. The default \code{n_tokens}, \code{n_types} and
\code{type_names} methods returns \code{NA}.
}
\description{
The \code{mclm} package contains a number of methods which don't do
anything useful at all, except from producing a warning, either because useful
implementations are yet to be implemented or because we simply couldn't think
of a useful implementation.

Typical examples are (i) defaults of generic methods that are only
meaningful for a limited number of classes, e.g. \code{n_tokens} and
\code{n_types}, and (ii) specific instances of generic methods which target classes
for which there is no implementation of the method (yet), e.g.
certain instances of the plotting method \code{\link{plot}}.
The methods listed on this page have in common that they
don't do anything substantial
(e.g. they don't return a meaningful number or they don't produce a plot)
and warn the user about this.
The methods listed on this page are implemented in order to prevent
cryptic error messages or warning from being show when the user
(accidentally) calls any of these methods.
}
\examples{
plot(re("^.{3,}"))

plot(summary(re("^.{3,}")))

n_tokens(1:10)

n_tokens(as_types(c("the", "a", "it")))

n_types(1:10)

n_types(as_tokens(c("a", "tiny", "example")))

type_names(1:10)
}
