% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{summary.types}
\alias{summary.types}
\alias{print.summary.types}
\title{Succinct Description of a 'types' Object}
\usage{
\method{summary}{types}(object, ...)

\method{print}{summary.types}(x, ...)
}
\arguments{
\item{object}{An object of class \code{types}.}

\item{...}{Additional arguments.}

\item{x}{An object of class \code{summary.types}.}
}
\value{
An object of class \code{summary.types}.
}
\description{
Build and/or print an object of the class \code{summary.types}.
}
\examples{

tps <- as_types(c("the", "a", "it", "is", "was", "are", "be", "being"))
summary(tps) 
print(summary(tps))
(tps_sum <- summary(tps))
names(tps_sum)
tps_sum[["n_types"]]
tps_sum$n_types
}
