% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqlist.R
\name{type_names.freqlist}
\alias{type_names.freqlist}
\title{Give Names of Types Represented in a Frequency List}
\usage{
\method{type_names}{freqlist}(x, ...)
}
\arguments{
\item{x}{Object of class \code{x}.}

\item{...}{Additional arguments.}
}
\value{
Character vector.
}
\description{
Return the names of the types represented in \code{x}.
}
\examples{
(flist <- freqlist("The man and the mouse.", as_text = TRUE))

n_types(flist)
type_names(flist)
type_freqs(flist)
}
