% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{write_types}
\alias{write_types}
\title{Write a vector of types to a text file}
\usage{
write_types(x, file, make_config_file = TRUE, ...)
}
\arguments{
\item{x}{Object of class \code{\link{types}}.}

\item{file}{Name of the output file}

\item{make_config_file}{Boolean value. Whether or not, next to the actual
types file, a second output file should be created containing a brief
description of the format of the types file.}

\item{...}{Additional arguments (not implemented).}
}
\value{
Invisibly, \code{x}.
}
\description{
This function writes an object of the class \code{\link{types}} to a text file. Each type
is written to a separate line. The file encoding that is used is
\code{"UTF-8"}.
}
\examples{
\dontrun{
  types <- as_types(c("first", "second", "third"))
  write_types(types, "file_with_types.txt")
  types_2 <- read_types("file_with_types.txt")
  }
}
\seealso{
\code{\link[=read_types]{read_types()}}
}
